Rename typedb.hpp to typedb_driver.hpp
